+++ ai_snake_lab/constants/DModelLRNN.py
@@ -17,6 +17,6 @@ class DModelRNN(ConstGroup):
     LEARNING_RATE: float = 0.0007  # Default: Learning rate
     HIDDEN_SIZE: int = 200  # Default: The number of nodes in the hidden layer
     RNN_LAYERS: int = 4  # Default: Number of RNN layers
-    P_VALUE: float = 0.2  # Default: Dropout layer value - 20%
+    P_VALUE: float = 0.18  # Default: Dropout layer value - 20%
     MODEL: str = "rnn"
     NON_LINEARITY: str = "tanh"